+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 4.13.3.0.0
+- Support new Tapjoy TJPrivacyPolicy APIs.
+- Drop support for deprecated Tapjoy mediation and video delegate APIs.
+- Show Tapjoy SDK initialization errors in console logs.
+- This version of the adapter has been certified with TapjoySDK 13.3.0.
+
 ### 4.13.0.0.1
 - Prevent multiple fullscreen loads for the same placement from happening concurrently.
 - This version of the adapters has been certified with TapjoySDK 13.0.0